requirements.txt
# Python dependencies for Residual_Carbon - Biochar Suitability Mapping Tool
# For detailed installation instructions, see README.md

# Core data processing
pandas>=2.0.0
numpy>=1.24.0

# Google Earth Engine
earthengine-api>=0.1.375

# Google Drive API (for automated download)
pydrive2>=1.16.0  # For automated download from Google Drive
google-auth>=2.23.0
google-auth-oauthlib>=1.1.0
google-auth-httplib2>=0.1.1

# Geospatial processing
# NOTE: On Windows, these are easier to install via Conda:
# conda install -c conda-forge geopandas rasterio shapely fiona pyproj gdal
geopandas>=0.14.0
shapely>=2.0.0
rasterio>=1.3.0
fiona>=1.9.0  # Required by geopandas
pyproj>=3.6.0  # Required by geopandas

# Spatial indexing
h3>=3.7.0

# Visualization
matplotlib>=3.7.0
folium>=0.14.0
pydeck>=0.8.0

# Configuration
pyyaml>=6.0

# Utilities
python-dotenv>=1.0.0

# Web browser automation (for auto-opening HTML)
webbrowser  # Built-in Python module

